minor issue solved
